implement task type list route, move many routes to ocs
