+++ appinfo/info.xml
@@ -6,14 +6,15 @@
 	<description><![CDATA[
 This app brings a user interface to use the Nextcloud text processing feature.
 
-It allows users to launch text processing tasks, be notified when they finish and see the results.
+It allows users to launch AI tasks, be notified when they finish and see the results.
 The assistant also appears in others apps like Text to easily process parts of a document.
 
 ### How to use it
 
-A new right header  menu entry appears. Once clicked, the assistant is displayed and you can select and task type and
-set the input text you want to process.
+A new right header menu entry appears. Once clicked, the assistant is displayed and you can select and task type and
+set the input you want to process.
 
+The task might run immediately or be scheduled depending on the time estimation given by the AI provider.
 Once a task is scheduled, it will run as a background job. When it is finished, you will receive a notification
 from which the results can be displayed.
 
@@ -21,10 +22,13 @@ Other apps can integrate with the assistant. For example, Text will display an i
 to directly select a task type to process this paragraph. Selecting a task this way will open the assistant with the task
 being pre-selected and the input text set.
 
-### Text processing providers
+## Features
 
 In the assistant, the list of available tasks depends on the available providers installed via other apps.
-This means you have complete freedom over which service/software will actually run your text processing tasks.
+This means you have complete freedom over which service/software will actually run your AI tasks.
+
+### Text processing providers
+
 So far, the [Large language model](https://github.com/nextcloud/llm#readme)
 and the [OpenAi/LocalAI integration](https://apps.nextcloud.com/apps/integration_openai) apps
 include text processing providers to:
@@ -33,6 +37,18 @@ include text processing providers to:
 * Generate a headline
 * Get an answer from a free prompt
 * Reformulate (OpenAi/LocalAi only)
+
+### Text to image (Image generation)
+
+Known providers:
+* [OpenAi/LocalAI integration](https://apps.nextcloud.com/apps/integration_openai)
+* [Text2Image Stable Diffusion](https://apps.nextcloud.com/apps/text2image_stablediffusion)
+
+### Speech to text (Audio transcription)
+
+Known providers:
+* [OpenAi/LocalAI integration](https://apps.nextcloud.com/apps/integration_openai)
+* [Local Whisper Speech-To-Text](https://apps.nextcloud.com/apps/stt_whisper)
 ]]>	</description>
 	<version>1.0.5</version>
 	<licence>agpl</licence>
