+++ lib/Controller/SpeechToTextController.php
@@ -23,39 +23,24 @@
 namespace OCA\TpAssistant\Controller;
 
 use Exception;
-use InvalidArgumentException;
 use OCA\TpAssistant\AppInfo\Application;
-use OCA\TpAssistant\Db\MetaTask;
-use OCA\TpAssistant\Db\MetaTaskMapper;
 use OCA\TpAssistant\Service\SpeechToText\SpeechToTextService;
 use OCP\AppFramework\Controller;
-use OCP\AppFramework\Db\DoesNotExistException;
-use OCP\AppFramework\Db\MultipleObjectsReturnedException;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
-use OCP\AppFramework\Http\DataResponse;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Services\IInitialState;
-use OCP\Files\NotFoundException;
-use OCP\Files\NotPermittedException;
-use OCP\IL10N;
 use OCP\IRequest;
-use OCP\PreConditionNotMetException;
-use Psr\Log\LoggerInterface;
-use RuntimeException;
 
 class SpeechToTextController extends Controller {
 
 	public function __construct(
-		string                      $appName,
-		IRequest                    $request,
+		string $appName,
+		IRequest $request,
 		private SpeechToTextService $service,
-		private LoggerInterface     $logger,
-		private IL10N               $l10n,
-		private IInitialState       $initialState,
-		private ?string             $userId,
-		private MetaTaskMapper      $metaTaskMapper,
+		private IInitialState $initialState,
+		private ?string $userId,
 	) {
 		parent::__construct($appName, $request);
 	}
@@ -67,10 +52,13 @@ class SpeechToTextController extends Controller {
 	#[NoAdminRequired]
 	#[NoCSRFRequired]
 	public function getResultPage(int $metaTaskId): TemplateResponse {
+		if ($this->userId === null) {
+			return new TemplateResponse('', '403', [], TemplateResponse::RENDER_AS_ERROR, Http::STATUS_FORBIDDEN);
+		}
 		$response = new TemplateResponse(Application::APP_ID, 'speechToTextResultPage');
 		try {
 			$initData = [
-				'task' => $this->internalGetTask($metaTaskId),
+				'task' => $this->service->internalGetTask($this->userId, $metaTaskId),
 			];
 		} catch (Exception $e) {
 			$initData = [
@@ -82,147 +70,5 @@ class SpeechToTextController extends Controller {
 		}
 		$this->initialState->provideInitialState('plain-text-result', $initData);
 		return $response;
-	}
-
-	/**
-	 * @param int $id Transcript ID
-	 * @return DataResponse
-	 */
-	#[NoAdminRequired]
-	public function getTranscript(int $id): DataResponse {
-		try {
-			return new DataResponse($this->internalGetTask($id)->getOutput());
-		} catch (Exception $e) {
-			return new DataResponse($e->getMessage(), intval($e->getCode()));
-		}
-	}
-
-	/**
-	 * Internal function to get transcription assistant tasks based on the assistant meta task id
-	 *
-	 * @param integer $id
-	 * @return MetaTask
-	 * @throws Exception
-	 */
-	private function internalGetTask(int $id): MetaTask {
-		try {
-			$metaTask = $this->metaTaskMapper->getUserMetaTask($id, $this->userId);
-
-			if($metaTask->getCategory() !== Application::TASK_CATEGORY_SPEECH_TO_TEXT) {
-				throw new Exception('Task is not a speech to text task.', Http::STATUS_BAD_REQUEST);
-			}
-
-			return $metaTask;
-		} catch (MultipleObjectsReturnedException $e) {
-			$this->logger->error('Multiple tasks found for one id: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			throw new Exception($this->l10n->t('Multiple tasks found'), Http::STATUS_BAD_REQUEST);
-		} catch (DoesNotExistException $e) {
-			throw new Exception($this->l10n->t('Transcript not found'), Http::STATUS_NOT_FOUND);
-		} catch (Exception $e) {
-			$this->logger->error('Error: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			throw new Exception(
-				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR,
-			);
-		}
-	}
-
-	/**
-	 * @return DataResponse
-	 * @throws NotPermittedException
-	 */
-	#[NoAdminRequired]
-	public function transcribeAudio(): DataResponse {
-		$audioData = $this->request->getUploadedFile('audioData');
-
-		if ($audioData['error'] !== 0) {
-			return new DataResponse('Error in audio file upload: ' . $audioData['error'], Http::STATUS_BAD_REQUEST);
-		}
-
-		if (empty($audioData)) {
-			return new DataResponse('Invalid audio data received', Http::STATUS_BAD_REQUEST);
-		}
-
-		if ($audioData['type'] !== 'audio/mp3' && $audioData['type'] !== 'audio/mpeg') {
-			return new DataResponse('Audio file must be in MP3 format', Http::STATUS_BAD_REQUEST);
-		}
-
-		try {
-			$this->service->transcribeAudio($audioData['tmp_name'], $this->userId);
-			return new DataResponse('ok');
-		} catch (RuntimeException $e) {
-			$this->logger->error(
-				'Runtime exception: ' . $e->getMessage(),
-				['app' => Application::APP_ID]
-			);
-			return new DataResponse(
-				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR
-			);
-		} catch (PreConditionNotMetException $e) {
-			$this->logger->error('No Speech-to-Text provider found: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			return new DataResponse(
-				$this->l10n->t('No Speech-to-Text provider found, install one from the app store to use this feature.'),
-				Http::STATUS_BAD_REQUEST
-			);
-		} catch (InvalidArgumentException $e) {
-			$this->logger->error('InvalidArgumentException: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			return new DataResponse(
-				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR
-			);
-		}
-	}
-
-	/**
-	 * @param string $path Nextcloud file path
-	 * @return DataResponse
-	 */
-	#[NoAdminRequired]
-	public function transcribeFile(string $path): DataResponse {
-		if ($path === '') {
-			return new DataResponse('Empty file path received', Http::STATUS_BAD_REQUEST);
-		}
-
-		try {
-			$this->service->transcribeFile($path, $this->userId);
-			return new DataResponse('ok');
-		} catch (NotFoundException $e) {
-			$this->logger->error('Audio file not found: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			return new DataResponse(
-				$this->l10n->t('Audio file not found.'),
-				Http::STATUS_NOT_FOUND
-			);
-		} catch (RuntimeException $e) {
-			$this->logger->error(
-				'Runtime exception: ' . $e->getMessage(),
-				['app' => Application::APP_ID]
-			);
-			return new DataResponse(
-				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR
-			);
-		} catch (NotPermittedException $e) {
-			$this->logger->error(
-				'No permission to create recording file/directory: ' . $e->getMessage(),
-				['app' => Application::APP_ID]
-			);
-			return new DataResponse(
-				$this->l10n->t('No permission to create recording file/directory, contact your sysadmin to resolve this issue.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR
-			);
-		} catch (PreConditionNotMetException $e) {
-			$this->logger->error('No Speech-to-Text provider found: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			return new DataResponse(
-				$this->l10n->t('No Speech-to-Text provider found, install one from the app store to use this feature.'),
-				Http::STATUS_BAD_REQUEST
-			);
-		} catch (InvalidArgumentException $e) {
-			$this->logger->error('InvalidArgumentException: ' . $e->getMessage(), ['app' => Application::APP_ID]);
-			return new DataResponse(
-				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
-				Http::STATUS_INTERNAL_SERVER_ERROR
-			);
-		}
 	}
 }
