+++ lib/Service/SpeechToText/SpeechToTextService.php
@@ -23,17 +23,24 @@
 namespace OCA\TpAssistant\Service\SpeechToText;
 
 use DateTime;
+use Exception;
 use InvalidArgumentException;
 use OCA\TpAssistant\AppInfo\Application;
+use OCA\TpAssistant\Db\MetaTask;
 use OCA\TpAssistant\Db\MetaTaskMapper;
+use OCP\AppFramework\Db\DoesNotExistException;
+use OCP\AppFramework\Db\MultipleObjectsReturnedException;
+use OCP\AppFramework\Http;
 use OCP\Files\File;
 use OCP\Files\Folder;
 use OCP\Files\IRootFolder;
 use OCP\Files\NotFoundException;
 use OCP\Files\NotPermittedException;
 use OCP\IConfig;
+use OCP\IL10N;
 use OCP\PreConditionNotMetException;
 use OCP\SpeechToText\ISpeechToTextManager;
+use Psr\Log\LoggerInterface;
 use RuntimeException;
 
 class SpeechToTextService {
@@ -43,7 +50,40 @@ class SpeechToTextService {
 		private IRootFolder $rootFolder,
 		private IConfig $config,
 		private MetaTaskMapper $metaTaskMapper,
+		private IL10N $l10n,
+		private LoggerInterface $logger,
 	) {
+	}
+
+	/**
+	 * Internal function to get transcription assistant tasks based on the assistant meta task id
+	 *
+	 * @param string $userId
+	 * @param integer $metaTaskId
+	 * @return MetaTask
+	 * @throws Exception
+	 */
+	public function internalGetTask(string $userId, int $metaTaskId): MetaTask {
+		try {
+			$metaTask = $this->metaTaskMapper->getUserMetaTask($metaTaskId, $userId);
+
+			if($metaTask->getCategory() !== Application::TASK_CATEGORY_SPEECH_TO_TEXT) {
+				throw new Exception('Task is not a speech to text task.', Http::STATUS_BAD_REQUEST);
+			}
+
+			return $metaTask;
+		} catch (MultipleObjectsReturnedException $e) {
+			$this->logger->error('Multiple tasks found for one id: ' . $e->getMessage(), ['app' => Application::APP_ID]);
+			throw new Exception($this->l10n->t('Multiple tasks found'), Http::STATUS_BAD_REQUEST);
+		} catch (DoesNotExistException $e) {
+			throw new Exception($this->l10n->t('Transcript not found'), Http::STATUS_NOT_FOUND);
+		} catch (Exception $e) {
+			$this->logger->error('Error: ' . $e->getMessage(), ['app' => Application::APP_ID]);
+			throw new Exception(
+				$this->l10n->t('Some internal error occurred. Contact your sysadmin for more info.'),
+				Http::STATUS_INTERNAL_SERVER_ERROR,
+			);
+		}
 	}
 
 	/**
